+++ benchtmpl/workflow/state.py
@@ -18,12 +18,26 @@ from datetime import datetime
 
 import traceback
 
+from benchtmpl.workflow.resource.base import ResourceDescriptor
+
+import benchtmpl.util.core as util
+
 
 """Definition of state type identifier."""
 STATE_ERROR = 'ERROR'
 STATE_PENDING = 'PENDING'
 STATE_RUNNING = 'RUNNING'
 STATE_SUCCESS = 'SUCCESS'
+
+
+"""Labels for serialization."""
+LABEL_CREATED_AT = 'createdAt'
+LABEL_FINISHED_AT = 'finishedAt'
+LABEL_MESSAGES = 'messages'
+LABEL_RESOURCES = 'resources'
+LABEL_STARTED_AT = 'startedAt'
+LABEL_STATE_TYPE = 'type'
+LABEL_STOPPED_AT = 'stoppedAt'
 
 
 class WorkflowState(object):
@@ -42,6 +56,37 @@ class WorkflowState(object):
         """
         self.type_id = type_id
         self.created_at = created_at if not created_at is None else datetime.now()
+
+    @staticmethod
+    def from_dict(doc):
+        """Get instance of a workflow state from a dictionary serialization as
+        created by the to_dict() method of the sub-types.
+
+        Parameters
+        ----------
+        doc: dict
+            Dictionary serialization of a workflow state
+
+        Returns
+        -------
+        benchtmpl.workflow.state.WorkflowState
+
+        Raises
+        ------
+        KeyError
+        ValueError
+        """
+        type_id = doc[LABEL_STATE_TYPE]
+        if type_id == STATE_PENDING:
+            return StatePending.from_dict(doc)
+        elif type_id == STATE_RUNNING:
+            return StateRunning.from_dict(doc)
+        elif type_id == STATE_ERROR:
+            return StateError.from_dict(doc)
+        elif type_id == STATE_SUCCESS:
+            return StateSuccess.from_dict(doc)
+        else:
+            raise ValueError('invalid state type \'{}\''.format(type_id))
 
     def is_active(self):
         """A workflow is in active state if it is either pending or running.
@@ -88,6 +133,18 @@ class WorkflowState(object):
         """
         return self.type_id == STATE_SUCCESS
 
+    def to_dict(self):
+        """Get dictionary serialization for the workflow state.
+
+        Returns
+        -------
+        dict
+        """
+        return {
+            LABEL_STATE_TYPE: self.type_id,
+            LABEL_CREATED_AT: self.created_at.isoformat()
+        }
+
 
 class StateError(WorkflowState):
     """Error state representation for a workflow run. The workflow has three
@@ -119,6 +176,46 @@ class StateError(WorkflowState):
         self.stopped_at = stopped_at if not stopped_at is None else datetime.now()
         self.messages = messages if not messages is None else list()
 
+    @staticmethod
+    def from_dict(doc):
+        """Get instance of a error workflow state from a dictionary
+        serialization.
+
+        Parameters
+        ----------
+        doc: dict
+            Dictionary serialization of a error workflow state
+
+        Returns
+        -------
+        benchtmpl.workflow.state.StateError
+
+        Raises
+        ------
+        KeyError
+        """
+        return StateError(
+            created_at=util.to_datetime(doc[LABEL_CREATED_AT]),
+            started_at=util.to_datetime(doc[LABEL_STARTED_AT]),
+            stopped_at=util.to_datetime(doc[LABEL_STOPPED_AT]),
+            messages=doc[LABEL_MESSAGES]
+        )
+
+    def to_dict(self):
+        """Get dictionary serialization for the workflow state. Extends the base
+        serialization with start and stop timestamps, and potential error
+        messages.
+
+        Returns
+        -------
+        dict
+        """
+        obj = super(StateError, self).to_dict()
+        obj[LABEL_STARTED_AT] = self.started_at.isoformat()
+        obj[LABEL_STOPPED_AT] = self.stopped_at.isoformat()
+        obj[LABEL_MESSAGES] = self.messages
+        return obj
+
 
 class StatePending(WorkflowState):
     """State representation for a pending workflow that is waiting to start
@@ -137,6 +234,26 @@ class StatePending(WorkflowState):
             type_id=STATE_PENDING,
             created_at=created_at
         )
+
+    @staticmethod
+    def from_dict(doc):
+        """Get instance of a pending workflow state from a dictionary
+        serialization.
+
+        Parameters
+        ----------
+        doc: dict
+            Dictionary serialization of a pending workflow state
+
+        Returns
+        -------
+        benchtmpl.workflow.state.StatePending
+
+        Raises
+        ------
+        KeyError
+        """
+        return StatePending(created_at=util.to_datetime(doc[LABEL_CREATED_AT]))
 
     def start(self):
         """Get instance of running state with the same create at timestamp as
@@ -190,12 +307,35 @@ class StateRunning(WorkflowState):
             messages=messages
         )
 
+    @staticmethod
+    def from_dict(doc):
+        """Get instance of a running workflow state from a dictionary
+        serialization.
+
+        Parameters
+        ----------
+        doc: dict
+            Dictionary serialization of a running workflow state
+
+        Returns
+        -------
+        benchtmpl.workflow.state.StateRunning
+
+        Raises
+        ------
+        KeyError
+        """
+        return StateRunning(
+            created_at=util.to_datetime(doc[LABEL_CREATED_AT]),
+            started_at=util.to_datetime(doc[LABEL_STARTED_AT])
+        )
+
     def success(self, resources=None):
         """Get instance of success state for a competed wokflow.
 
         Parameters
         ----------
-        resources: dict(benchtmpl.workflow.resource.ResourceHandle), optional
+        resources: dict(benchtmpl.workflow.resource.ResourceDescriptor), optional
             Optional dictionary of created resources
 
         Returns
@@ -207,6 +347,18 @@ class StateRunning(WorkflowState):
             started_at=self.started_at,
             resources=resources
         )
+
+    def to_dict(self):
+        """Get dictionary serialization for the workflow state. Extends the base
+        serialization with the start timestamp.
+
+        Returns
+        -------
+        dict
+        """
+        obj = super(StateRunning, self).to_dict()
+        obj[LABEL_STARTED_AT] = self.started_at.isoformat()
+        return obj
 
 
 class StateSuccess(WorkflowState):
@@ -227,8 +379,12 @@ class StateSuccess(WorkflowState):
             Timestamp when the workflow started running
         finished_at: datetime.datetime, optional
             Timestamp when workflow execution completed
-        resources: dict(benchtmpl.workflow.resource.ResourceHandle), optional
+        resources: dict(benchtmpl.workflow.resource.ResourceDescriptor), optional
             Optional dictionary of created resources
+
+        Raises
+        ------
+        ValueError
         """
         super(StateSuccess, self).__init__(
             type_id=STATE_SUCCESS,
@@ -236,4 +392,59 @@ class StateSuccess(WorkflowState):
         )
         self.started_at = started_at
         self.finished_at = finished_at if not finished_at is None else datetime.now()
-        self.resources = resources if not resources is None else dict()
+        if not resources is None:
+            if isinstance(resources, dict):
+                self.resources = resources
+            elif isinstance(resources, list):
+                self.resources = dict()
+                for res in resources:
+                    self.resources[res.identifier] = res
+            else:
+                raise ValueError('invalid data type for resources')
+        else:
+            self.resources = dict()
+
+    @staticmethod
+    def from_dict(doc):
+        """Get instance of a success workflow state from a dictionary
+        serialization.
+
+        Parameters
+        ----------
+        doc: dict
+            Dictionary serialization of a success workflow state
+
+        Returns
+        -------
+        benchtmpl.workflow.state.StateSuccess
+
+        Raises
+        ------
+        KeyError
+        """
+        resources = dict()
+        if LABEL_RESOURCES in doc:
+            for res in doc[LABEL_RESOURCES]:
+                desc = ResourceDescriptor.from_dict(res)
+                resources[desc.identifier] = desc
+        return StateSuccess(
+            created_at=util.to_datetime(doc[LABEL_CREATED_AT]),
+            started_at=util.to_datetime(doc[LABEL_STARTED_AT]),
+            finished_at=util.to_datetime(doc[LABEL_FINISHED_AT]),
+            resources=resources
+        )
+
+    def to_dict(self):
+        """Get dictionary serialization for the workflow state. Extends the base
+        serialization with start and finished timestamps, and potential result
+        resources.
+
+        Returns
+        -------
+        dict
+        """
+        obj = super(StateSuccess, self).to_dict()
+        obj[LABEL_STARTED_AT] = self.started_at.isoformat()
+        obj[LABEL_FINISHED_AT] = self.finished_at.isoformat()
+        obj[LABEL_RESOURCES] = [r.to_dict() for r in self.resources.values()]
+        return obj