+++ changelog.md
@@ -8,3 +8,9 @@
 ### 0.1.1 - 2019-06-21
 
 * Move multi-process backend to separate repository
+
+
+### 0.1.2 - 2019-06-23
+
+* Add (de-)serializer methods for workflow state classes
+* Change behavior of resource descriptor serialization